backend/requirements.txt
fastapi
uvicorn
websockets
adbutils
tidevice
loguru
psutil
aiofiles
Pillow
numpy
aiomysql
python-dotenv
passlib
bcrypt
PyJWT
pydantic
pydantic-settings
email-validator
PyYAML
redis
